# docs/openapi/evaluator.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Oscanner Evaluator API",
    "summary": "Contributor capability evaluation service",
    "description": "Extract GitHub and Gitee activity, discover contributor identities, evaluate contributors with plugin-defined rubrics, analyze trajectories, and proxy repository runner workflows.",
    "license": {
      "name": "Apache-2.0"
    },
    "version": "0.1.6"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local evaluator service"
    }
  ],
  "paths": {
...
    "/api/benchmark/repos": {
      "get": {
        "tags": [
          "benchmark"
        ],
        "summary": "Get Benchmark Repos",
        "description": "Get paginated list of benchmark repos.",
        "operationId": "get_benchmark_repos_api_benchmark_repos_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Per Page"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/runner/{path}": {
      "get": {
        "tags": [
          "runner"
        ],
        "summary": "Proxy runner GET request",
        "description": "Proxy all /api/runner/* requests to the repos_runner service\n\nArgs:\n    path: The path after /api/runner/\n    request: The incoming request\n\nReturns:\n    Proxied response from repos_runner service",
        "operationId": "proxy_runner_get_api_runner_path",
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "runner"
        ],
        "summary": "Proxy runner POST request",
        "description": "Proxy all /api/runner/* requests to the repos_runner service\n\nArgs:\n    path: The path after /api/runner/\n    request: The incoming request\n\nReturns:\n    Proxied response from repos_runner service",
        "operationId": "proxy_runner_post_api_runner_path",
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "runner"
        ],
        "summary": "Proxy runner PUT request",
        "description": "Proxy all /api/runner/* requests to the repos_runner service\n\nArgs:\n    path: The path after /api/runner/\n    request: The incoming request\n\nReturns:\n    Proxied response from repos_runner service",
        "operationId": "proxy_runner_put_api_runner_path",
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "runner"
        ],
        "summary": "Proxy runner PATCH request",
        "description": "Proxy all /api/runner/* requests to the repos_runner service\n\nArgs:\n    path: The path after /api/runner/\n    request: The incoming request\n\nReturns:\n    Proxied response from repos_runner service",
        "operationId": "proxy_runner_patch_api_runner_path",
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "runner"
        ],
        "summary": "Proxy runner DELETE request",
        "description": "Proxy all /api/runner/* requests to the repos_runner service\n\nArgs:\n    path: The path after /api/runner/\n    request: The incoming request\n\nReturns:\n    Proxied response from repos_runner service",
        "operationId": "proxy_runner_delete_api_runner_path",
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/checkers/run": {
      "post": {
        "tags": [
          "checkers"
        ],
        "summary": "Run Checker",
        "description": "Run a checker on a specific commit.\n\nArgs:\n    request: CheckerRunRequest with checker_id, platform, owner, repo, commit_sha, and optional files\n    \nReturns:\n    Checker execution result (dict with success, score, passed, total, details, message, etc.)",
        "operationId": "run_checker_api_checkers_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckerRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Run Checker Api Checkers Run Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CheckerRunRequest": {
        "properties": {
          "checker_id": {
            "type": "string",
            "title": "Checker Id"
          },
          "platform": {
            "type": "string",
            "title": "Platform"
          },
          "owner": {
            "type": "string",
            "title": "Owner"
          },
          "repo": {
            "type": "string",
            "title": "Repo"
          },
          "commit_sha": {
            "type": "string",
            "title": "Commit Sha"
          },
          "files": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Files"
          },
          "worktree_base": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Worktree Base"
          }
        },
        "type": "object",
        "required": [
          "checker_id",
          "platform",
          "owner",
          "repo",
          "commit_sha"
        ],
        "title": "CheckerRunRequest",
        "description": "Request model for running a checker."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}